move add_vary_to_headerlist into pyramid.response
